+++ CHANGELOG.md
@@ -4,6 +4,27 @@ All notable changes to `package-sync` will be documented in this file.
 
 ---
 
+=======
+## 1.3.0 - unreleased
+
+- fix `dist/package-sync` script to work better with `npm` and `npx`
+
+## 1.2.0 - 2021-04-08
+
+- add config options to define the values used when replacing template variables
+
+## 1.1.0 - 2021-04-08
+
+- add config option to define the package vendor name
+- use package vendor name when pulling packages
+- add `--config <filename>` flag to `analyze` and `fix` commands
+- minor fixes
+
+## 1.0.1 - 2021-04-08
+
+- change npm package name
+- fix npm bin script
+
 ## 1.0.0 - 2021-04-08
 
 - initial release